[impr-#97] Set default qty of 1 when creating new orders manually
If add qty field is empty set a default vallue of 1

Solution is maybe not the most elegant but makes sure we don't overwrite existing values.
Only if the fields is empty set a default value. A an improvement one could get the value from the
mimimal_order_qty settings in settings (this is the way you set a default order qty value in frontend instead
of leaving it empty)

Please imagine use entering orders every day via adminhtml and every every time this fields is empty

A starting value of 1 already solves 95% of these cases