[impr-#101] Only use base url for homepage in generated sitemap.xml
tja. works for 99% is important to have right for 100%. But this solution is not sound proof.
Still I would advise to accept it. As de base-url is always the homepage ... and I am not sure
we would want to create another config option with "use baseurl als main homepage yes/no"