[bug-#297] Only set email_sent if customer was actually notified
In all other places where a "email" flag is used, email_sent is only set if the flag is true.

PS:
In my opinion, the $notifiyCustomer condition is a better solution than assigning
$notifiyCustomer directly to "email_sent". One can send invoices/creditmemos multiple
times. Just a second run could make it look like an email was never sent
(even though it was sent in the first run).