[impr-#197] Support progressive jpegs
Disabled per default because it takes a bit longer to process progressive.
For a single image this is not a big deal but if you invalidate your
entire image cache by changing something like the watermark and you are in a
high-traffic scenario then it can really stress the server.

Specify a non-zero size in megapixels to enable progressive image compression via
`catalog/product_image/progressive_threshold`. Smaller images will not use progressive
compression. E.g. 0.5 will enable progressive mode for images equal to or larger than
half a megapixel.

refs #95