[bug-#500] Backend user is logged out in refreshAcl()
When the `reload_acl_flag` is set to 1  the `refreshAcl` method is setting
the flag value back to 0. However, it also unsets the password and in doing
so causes the `reload` method of the User model to fail the password check.
The id is then set to NULL and this gets saved in the users session so
the *next* page load results in the user being effectively logged out
based on how it is checking in the `isLoggedIn` method.