[bug-#978] vsprintf() called with no arguments in Mage_Core_Model_Translate::translate()
This patch fixes "vsprintf(): Too few arguments" notice, when magento tries to
prepare JS language labels in Mage_Core_Helper_Js
(they are translated without passing arguments, as args are replaced on js side).
This way we avoid  the notice and processing of the notice in the error handler
(which adds some performance penalty to every request).